add the redis config

+++ company/company-server/src/main/java/com/zts/dao/CompanyDao.java
@@ -5,6 +5,6 @@ import com.zts.entity.Company;
 /**
  * @author zhangtusheng
  */
-public interface CompanyDao extends IBaseDao<Company,Long>{
+public interface CompanyDao extends IBaseDao<Company,Integer>{
 
 }
